OCT-111: moved GetCatalogProductSelectionCriteriaQuery
