+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Catalog/GetCatalogProductSelectionCriteriaQueryTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence\Catalog;
 
-use Akeneo\Catalogs\Infrastructure\Persistence\GetCatalogProductSelectionCriteriaQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Catalog\GetCatalogProductSelectionCriteriaQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 use PHPUnit\Framework\Assert;
 
@@ -12,7 +12,7 @@ use PHPUnit\Framework\Assert;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
- * @covers \Akeneo\Catalogs\Infrastructure\Persistence\GetCatalogProductSelectionCriteriaQuery
+ * @covers \Akeneo\Catalogs\Infrastructure\Persistence\Catalog\GetCatalogProductSelectionCriteriaQuery
  */
 class GetCatalogProductSelectionCriteriaQueryTest extends IntegrationTestCase
 {
